Fixes
- add \n to module seeders
- fix academical hours
- change payment value from text to int
- add filter to instructor by category id
- remove possibility to delete exam module
- add false exam filter to exams in module

+++ app/Components/Payments/BusinessLayer/Read.php
@@ -131,7 +131,7 @@ class Read
             ->all();
         $dept = 0.0;
         foreach ($valuesArr as $value) {
-            $dept += $coursePrice - (float) $value;
+            $dept += $coursePrice - (int) $value;
         }
         return round($dept);
     }
